node: Store & retrieve node information via kvstore
Each agent stores its own node information in the kvstore for all other agents
to retrieve. Although nodes belong to a cluster, the association does not have
any effect yet, it is in preparation for later features.
u
The node structure defines how to reach the node, either via an encapsulation
or via a direct route. Encapsulation is always preferred unless explicitly
disabled. A node which does not announce reachability via either encapsulation
or direct routing will not be reachable.

THe kvstore based node detection becomes the new default, the existing k8s
annotation based code is preseved and can be enabled with the option
--k8s-use-node-annotations

Signed-off-by: Thomas Graf <thomas@cilium.io>