etcd: Run etcd version check in the background
When an etcd endpoint is unavailable, the version check retrieves the version
of each etcd endpoint. If one etcd endpoint is not available, the version check
will time out eventually. This is currently a blocking operation which delays
the renewal of etcd sessions. Run the version check in the background instead
to not delay renewal of etcd sessions.

Never return errors on version checks except when the version check clearly
fails.

Fixes: #3496

Signed-off-by: Thomas Graf <thomas@cilium.io>