pkg/policy: remove \"changed\" return value from regenerateConsumable
regenerateConsumable is only called within regeneratePolicy. The documentation
for regeneratePolicy dictates that it returns whether the BPF for the given
endpoint such be regenerated. However, regenerateConsumable returned \"changed\"
only when the PolicyMap for the endpoint / PolicyMaps for the Consumable were
updated. PolicyMap state changes do not necessitate regeneration of the BPF
program for an endpoint. However, the result of regenerateConsumable, which only
returned \"true\" for when changes were made to PolicyMaps, was used in
determining whether policy was changed in regeneratePolicy, and thus meant that
a change in the PolicyMaps resulted in regeneratePolicy specifying that BPF
needed to be regenerated for the given endpoint. As explained, since BPF
regeneration is not necessary in the case of PolicyMap updates, just remove the
return value for whether regenerateConsumable changed the PolicyMaps for
endpoints when determining whether BPF needs to be regenerated.

Signed-off by: Ian Vernon <ian@cilium.io>